+++ .gitignore
@@ -1,4 +1,4 @@
-node_modules
-out
-dist
-docs
+/node_modules
+/out
+/dist
+/docs
