+++ dist-overwrite/index.d.ts
@@ -1,0 +1,11 @@
+export * as log from "./log";
+export * as rules from "./rules";
+export * as items from "./items";
+export * as things from "./things";
+export * as metadata from "./metadata";
+export * as triggers from "./triggers";
+export * as actions from "./actions";
+export * as utils from "./utils";
+export * as osgi from "./osgi";
+export * as cache from "./cache";
+export * as time from "@js-joda/core";
