+++ dist-overwrite/items/index.d.ts
@@ -1,0 +1,3 @@
+export * from "./managed";
+export * from "./item";
+export * as provider from "./items-provider";
