+++ index2.ts
@@ -1,0 +1,13 @@
+// export * as log from "./log";
+// // export * as rules from "./rules"
+
+// // import * as rules from "./rules";
+// // export rules;
+
+// export * from "./rules.d.ts"
+
+// module.exports = {
+//     rules: require("/rules")
+// }
+
+module.exports.rules = require("/rules")
