+++ items/index.js
@@ -1,10 +1,11 @@
 /**
  * Items namespace.
  * This namespace handles querying and updating openHAB Items.
- * @namespace items
+ * namespace items
  */
 
 module.exports = {
     ...require('./managed'),
+    Item: require("./item"),
     provider: require('./items-provider')
 }
