+++ items/managed.js
@@ -1,12 +1,16 @@
-
+/**
+ * @typedef { import("../index").HostItem } HostItem
+ * @typedef { import("../index").HostGroupFunction } HostGroupFunction
+ * @typedef { import("../index").HostState } HostState
+ * @typedef { import("./item") } Item
+ * 
+*/
 
 const osgi = require('../osgi');
 const utils = require('../utils');
 const log = require('../log')('items');
-const metadata = require('../metadata');
-const ItemHistory = require('./item-history');
 
-const { UnDefType, events, itemRegistry } = require('@runtime');
+const { itemRegistry } = require('@runtime');
 
 const itemBuilderFactory = osgi.getService("org.openhab.core.items.ItemBuilderFactory");
 
@@ -21,227 +25,7 @@ const managedItemProvider = osgi.getService("org.openhab.core.items.ManagedItemP
  */
 const DYNAMIC_ITEM_TAG = "_DYNAMIC_";
 
-/**
- * Class representing an openHAB Item
- * 
- * @memberOf items
- */
-class Item {
-    /**
-     * Create an Item, wrapping a native Java openHAB Item. Don't use this constructor, instead call {@link getItem}.
-     * @param {HostItem} rawItem Java Item from Host
-     * @hideconstructor
-     */
-    constructor(rawItem) {
-        if (typeof rawItem === 'undefined') {
-            throw Error("Supplied item is undefined");
-        }
-        this.rawItem = rawItem;
 
-        /**
-         * Access historical states for this item
-         * @type {items.ItemHistory}
-         */
-        this.history = new ItemHistory(rawItem);
-    }
-
-    /**
-     * The type of the item: the Simple (without package) name of the Java item type, such as 'Switch'.
-     * @return {String} the type
-     */
-    get type() {
-        return this.rawItem.getClass().getSimpleName();
-    }
-
-    /**
-     * The name of the item.
-     * @return {String} the name
-     */
-    get name() {
-        return this.rawItem.getName();
-    }
-
-    /**
-     * The label attached to the item
-     * @return {String} the label
-     */
-    get label() {
-        return this.rawItem.getLabel();
-    }
-
-    /**
-     * The state of the item, as a string.
-     * @return {String} the item's state
-     */
-    get state() {
-        return this.rawState.toString();
-    }
-
-    /**
-     * The raw state of the item, as a java object.
-     * @return {HostState} the item's state
-     */
-    get rawState() {
-        return this.rawItem.state;
-    }
-
-    /**
-     * Members / children / direct descendents of the current group item (as returned by 'getMembers()'). Must be a group item.
-     * @returns {Item[]} member items
-     */
-    get members() {
-        return utils.javaSetToJsArray(this.rawItem.getMembers()).map(raw => new Item(raw));
-    }
-
-    /**
-     * All descendents of the current group item (as returned by 'getAllMembers()'). Must be a group item.
-     * @returns {Item[]} all descendent items
-     */
-    get descendents() {
-        return utils.javaSetToJsArray(this.rawItem.getAllMembers()).map(raw => new Item(raw));
-    }
-
-    /**
-     * Whether this item is initialized.
-     * @type {Boolean}
-     * @returns true iff the item has not been initialized
-     */
-    get isUninitialized() {
-        if (this.rawItem.state instanceof UnDefType
-            || this.rawItem.state.toString() == "Undefined"
-            || this.rawItem.state.toString() == "Uninitialized"
-        ) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     * Gets metadata values for this item.
-     * @param {String} namespace The namespace for the metadata to retreive
-     * @returns {String} the metadata associated with this item and namespace
-     */
-    getMetadataValue(namespace) {
-        return metadata.getValue(this.name, namespace);
-    }
-
-    /**
-     * Updates metadata values for this item.
-     * @param {String} namespace The namespace for the metadata to update
-     * @param {String} value the value to update the metadata to
-     * @returns {String} the updated value
-     */
-    updateMetadataValue(namespace, value) {
-        return metadata.updateValue(this.name, namespace, value);
-    }
-
-    /**
-     * Inserts or updates metadata values for this item.
-     * @param {String} namespace The namespace for the metadata to update
-     * @param {String} value the value to update the metadata to
-     * @returns {Boolean} true iff a new value was inserted
-     */
-    upsertMetadataValue(namespace, value) {
-        return metadata.upsertValue(this.name, namespace, value);
-    }
-
-    /**
-     * Updates metadata values for this item.
-     * @param {Map} namespaceToValues A map of namespaces to values to update
-     */
-    updateMetadataValues(namespaceToValues) {
-        for(let k in namespaceToValues) {
-            metadata.updateValue(this.name, k, namespaceToValues[k]);
-        }
-    }
-
-    /**
-     * Sends a command to the item
-     * @param {String|HostState} value the value of the command to send, such as 'ON'
-     * @see sendCommandIfDifferent
-     * @see postUpdate
-     */
-    sendCommand(value) {
-        log.debug("Sending command {} to {}", value, this.name);
-        events.sendCommand(this.rawItem, value);
-    }
-
-    /**
-     * Sends a command to the item, but only if the current state is not what is being sent.
-     * Note
-     * @param {String|HostState} value the value of the command to send, such as 'ON'
-     * @returns {Boolean} true if the command was sent, false otherwise
-     * @see sendCommand
-     */
-    sendCommandIfDifferent(value) {
-        if (value.toString() != this.state.toString()) {
-            this.sendCommand(value);
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * Posts an update to the item
-     * @param {String|HostState} value the value of the command to send, such as 'ON'
-     * @see sendCommand
-     */
-    postUpdate(value) {
-        events.postUpdate(this.rawItem, value);
-        log.debug("Posted update {} to {}", value, this.name);
-    }
-
-    /**
-     * Adds groups to this item
-     * @param {Array<String|Item>} groupNamesOrItems names of the groups (or the group items themselves)
-     */
-    addGroups(...groupNamesOrItems) {
-        let groupNames = groupNamesOrItems.map((x) => (typeof x === 'string') ? x : x.name);
-        this.rawItem.addGroupNames(groupNames);
-        managedItemProvider.update(this.rawItem);
-    }
-
-    /**
-     * Removes groups from this item
-     * @param {Array<String|Item>} groupNamesOrItems names of the groups (or the group items themselves)
-     */
-    removeGroups(...groupNamesOrItems) {
-        let groupNames = groupNamesOrItems.map((x) => (typeof x === 'string') ? x : x.name);
-        for(let groupName of groupNames) {
-            this.rawItem.removeGroupName(groupName);
-        }
-        managedItemProvider.update(this.rawItem);
-    }
-
-    /**
-     * Gets the tags from this item
-     */
-    get tags() {
-        return utils.javaSetToJsArray(this.rawItem.getTags());
-    }
-
-    /**
-     * Adds tags to this item
-     * @param {Array<String>} tagNames names of the tags to add
-     */
-    addTags(...tagNames) {
-        this.rawItem.addTags(tagNames);
-        managedItemProvider.update(this.rawItem);
-    }
-
-    /**
-     * Removes tags from this item
-     * @param {Array<String>} tagNames names of the tags to remove
-     */
-    removeTags(...tagNames) {
-        for(let tagName of tagNames) {
-            this.rawItem.removeTag(tagName);
-        }
-        managedItemProvider.update(this.rawItem);
-    }
-}
 
 /**
  * Creates a new item within OpenHab. This item is not registered with any provider.
@@ -258,16 +42,16 @@ class Item {
  * @param {String[]} [tags] an array of tags for the Item
  * @param {HostItem} [giBaseType] the group Item base type for the Item
  * @param {HostGroupFunction} [groupFunction] the group function used by the Item
- * @param {Map} [itemMetadata] a map of metadata to set on the item
+ * @param {Map<any, any>} [itemMetadata] a map of metadata to set on the item
  */
 const createItem = function (itemName, itemType, category, groups, label, tags, giBaseType, groupFunction, itemMetadata) {
     itemName = safeItemName(itemName);
-    
+
     let baseItem;
     if (itemType === 'Group' && typeof giBaseType !== 'undefined') {
         baseItem = itemBuilderFactory.newItemBuilder(giBaseType, itemName + "_baseItem").build()
     }
-    
+
     if (itemType !== 'Group') {
         groupFunction = undefined;
     }
@@ -407,15 +191,15 @@ const replaceItem = function (/* same args as addItem */) {
  * @memberOf items
  * @param {String} name the name of the item
  * @param {String} nullIfMissing whether to return null if the item cannot be found (default is to throw an exception)
- * @return {items.Item} the item
+ * @return {Item} the item
  */
 const getItem = (name, nullIfMissing = false) => {
     try {
         if (typeof name === 'string' || name instanceof String) {
             return new Item(itemRegistry.getItem(name));
         }
-    } catch(e) {
-        if(nullIfMissing) {
+    } catch (e) {
+        if (nullIfMissing) {
             return null;
         } else {
             throw e;
@@ -428,7 +212,7 @@ const getItem = (name, nullIfMissing = false) => {
  * 
  * @memberOf items
  * @param {String[]} tagNames an array of tags to match against
- * @return {items.Item[]} the items with a tag that is included in the passed tags
+ * @return {Item[]} the items with a tag that is included in the passed tags
  */
 const getItemsByTag = (...tagNames) => {
     return utils.javaSetToJsArray(itemRegistry.getItemsByTag(tagNames)).map(i => new Item(i));
@@ -441,8 +225,8 @@ const getItemsByTag = (...tagNames) => {
  * @returns {String} a valid item name
  */
 const safeItemName = s => s.
-        replace(/[\"\']/g, ''). //delete
-        replace(/[^a-zA-Z0-9]/g, '_'); //replace with underscore
+    replace(/[\"\']/g, ''). //delete
+    replace(/[^a-zA-Z0-9]/g, '_'); //replace with underscore
 
 module.exports = {
     safeItemName,
@@ -452,7 +236,7 @@ module.exports = {
     replaceItem,
     createItem,
     removeItem,
-    Item,
+    // Item,
     /**
      * Custom indexer, to allow static item lookup.
      * @example
