--- items/package.json
@@ -1,5 +1,0 @@
-{
-    "name": "items",
-    "version": "0.0.1",
-    "main": "items.js"
-}
