+++ metadata/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    ...require("./metadata"),
+    provider: require('./metadata-provider')
+}
