--- metadata/package.json
@@ -1,5 +1,0 @@
-{
-    "name": "metadata",
-    "version": "0.0.1",
-    "main": "metadata.js"
-}
