+++ package.json
@@ -8,9 +8,10 @@
         "type": "git",
         "url": "git://github.com/openhab/openhab-js.git"
     },
+    "types": "dist/index.d.ts",
     "dependencies": {
         "@js-joda/core": "^4.3.1",
-        "parse-duration": "^0.1.1"
+        "parse-duration": "^1.0.2"
     },
     "scripts": {
         "test": "mocha",
