+++ rules/index.js
@@ -1,4 +1,7 @@
-const rules = require('./rules');
-const builder = require('./rule-builder');
-
-module.exports = Object.assign(rules, builder);
+/**
+ * @namespace rules
+ */
+module.exports = {
+    ...require('./rulesx'),
+    ...require('./rule-builder')
+}
