--- rules/package.json
@@ -1,18 +1,0 @@
-{
-    "name": "rules",
-    "version": "0.0.1",
-    "description": "Rules API for openHAB",
-    "main": "index.js",
-    "dependencies": {
-        "cronstrue": "^1.84.0",
-        "mocha": "^6.2.3",
-        "parse-duration": "^0.1.1",
-        "proxyquire": "^2.1.3"
-    },
-    "scripts": {
-        "test": "mocha"
-    },
-    "devDependencies": {
-        "rewiremock": "^3.13.9"
-    }
-}
