--- things/package.json
@@ -1,5 +1,0 @@
-{
-    "name": "things",
-    "version": "0.0.1",
-    "main": "things.js"
-}
